Shift from `autocxx` to `cxx`
Rewrites the library to rely on `cxx` rather than `autocxx` as the
bindings back-end. Updates the building script accordingly.

Note that a custom version of cxx is used: it includes the patch
dtolnay/cxx#1715 for the `#[safe_shared_extern]` attribute.

Motivation for the change: wasm32 support (#8).

xgrammar-rs now compiles for `wasm32-wasip1` (and, supposedly, other
wasi targets), though for now it still requires the `WASI_SYSROOT`
environment variable to be set to the path to
[wasi-sdk](https://github.com/WebAssembly/wasi-sdk) version 33 or newer
compiled with exception support (`-DWASI_SDK_EXCEPTIONS=ON`).

+++ src/cxx_utils/grammar_compiler.hpp
@@ -17,9 +17,9 @@ namespace cxx_utils {
 
 inline std::unique_ptr<xgrammar::GrammarCompiler> make_grammar_compiler(
     const xgrammar::TokenizerInfo& tokenizer_info,
-    int max_threads,
+    int32_t max_threads,
     bool cache_enabled,
-    long long cache_limit_bytes,
+    int64_t cache_limit_bytes,
     std::string* error_out
 ) {
   try {
@@ -51,13 +51,13 @@ inline std::unique_ptr<xgrammar::CompiledGrammar> compiler_compile_json_schema(
     const std::string& schema,
     bool any_whitespace,
     bool has_indent,
-    int indent,
+    int32_t indent,
     bool has_separators,
     const std::string& separator_comma,
     const std::string& separator_colon,
     bool strict_mode,
     bool has_max_whitespace_cnt,
-    int max_whitespace_cnt,
+    int32_t max_whitespace_cnt,
     std::string* error_out
 ) {
   try {
